refactor(ChristmasPromotion): 돈을 다루므로 int가 아닌 long을 사용하도록 리팩토링

+++ src/main/java/christmas/domain/ChristmasPromotion.java
@@ -3,21 +3,21 @@ package christmas.domain;
 public class ChristmasPromotion {
     private static final int START_DAY = 1;
     private static final int END_DAY = 25;
-    private static final int MIN_TOTAL_PRICE = 10000;
-    private static final int DEFAULT_DISCOUNT = 1000;
-    private static final int INCREASING_DISCOUNT = 100;
-    private static final int NO_DISCOUNT = 0;
+    private static final long MIN_TOTAL_PRICE = 10000;
+    private static final long DEFAULT_DISCOUNT = 1000;
+    private static final long INCREASING_DISCOUNT = 100;
+    private static final long NO_DISCOUNT = 0;
 
     private final VisitDay visitDay;
-    private final int totalPrice;
+    private final long totalPrice;
 
-    public ChristmasPromotion(VisitDay visitDay, int totalPrice) {
+    public ChristmasPromotion(VisitDay visitDay, long totalPrice) {
         this.visitDay = visitDay;
         this.totalPrice = totalPrice;
     }
 
-    public int calculateDiscount() {
-        if (visitDay.isInChristmasPromotion() && totalPrice >= MIN_TOTAL_PRICE) {
+    public long calculateDiscount() {
+        if (visitDay.isInChristmasPromotion() && (totalPrice >= MIN_TOTAL_PRICE)) {
             return DEFAULT_DISCOUNT + (visitDay.getDifferenceFromMinDay() * INCREASING_DISCOUNT);
         }
         return NO_DISCOUNT;
